+++ qoala/sim/host/hostinterface.py
@@ -70,7 +70,7 @@ class HostInterface(ComponentProtocol):
             )
         )
 
-    def wait(self, delta_time: int) -> Generator[EventExpression, None, None]:
+    def wait(self, delta_time: float) -> Generator[EventExpression, None, None]:
         self._schedule_after(delta_time, EVENT_WAIT)
         event_expr = EventExpression(source=self, event_type=EVENT_WAIT)
         yield event_expr
